Lap Time Tracking and Plots
Fixed Best Lap tracking.
Now the best laps of each episode are saved in a list.

#Requires CsvHelper library
https://www.reddit.com/r/Unity3D/comments/6qg197/readingwriting_csvs_in_unity/

When the Testing is finished it saves a .csv file it in the LocalLow/DefaultCompany/CarsML-2021-main folder with their best lap times.

If the testing is in Heuristic mode then it saves them in the same path but first it creates a new folder called Heuristic-(Datetime) and adds the .csv files there.

Added Plots/plotBestLaptime.py . Requires matplotlib
https://matplotlib.org/stable/users/installing.html

It takes the data files from the NON HEURISTIC modes and plots them in (all together, side by side and each on its own) as well as saving the plot as a picture.

I'll be commiting the heuristic plots soon, in a new .py script